commander: update mission_result for mission check
This is an attempt to fix the race condition that can happen in SITL
testing where switching to mission mode happens right after mission
upload. What presumably happens is that commander has not updated the
mission_result yet and therefore denies to switch to mission mode.

To fix, we now update the mission_result topic just before checking the
mission validity. We still do the regular checks as before.